+++ crates/ruma-state-res/tests/it/resolve.rs
@@ -10,8 +10,9 @@ use std::{
 };
 
 use ruma_common::{
-    room_version_rules::AuthorizationRules, MilliSecondsSinceUnixEpoch, OwnedEventId, OwnedRoomId,
-    OwnedUserId, RoomId, RoomVersionId, UserId,
+    room_version_rules::{AuthorizationRules, StateResolutionV2Rules},
+    MilliSecondsSinceUnixEpoch, OwnedEventId, OwnedRoomId, OwnedUserId, RoomId, RoomVersionId,
+    UserId,
 };
 use ruma_events::{StateEventType, TimelineEventType};
 use ruma_state_res::{resolve, Event, StateMap};
@@ -207,20 +208,32 @@ fn test_resolve(paths: &[&str]) -> Snapshots {
             .expect("the m.room.create PDU's content should be valid")
             .room_version
     };
-    let rules = room_version_id.rules().expect("room version should be supported").authorization;
+    let rules = room_version_id.rules().expect("room version should be supported");
+    let auth_rules = rules.authorization;
+    let state_res_rules =
+        rules.state_res.v2_rules().expect("resolve only supports state resolution version 2");
 
     // Resolve PDUs iteratively, using the ordering of `prev_events`.
-    let iteratively_resolved_state =
-        resolve_iteratively(&rules, pdu_batches.iter().flat_map(|x| x.iter()))
-            .expect("iterative state resolution should succeed");
+    let iteratively_resolved_state = resolve_iteratively(
+        &auth_rules,
+        &state_res_rules,
+        pdu_batches.iter().flat_map(|x| x.iter()),
+    )
+    .expect("iterative state resolution should succeed");
 
     // Resolve PDUs in batches by file
     let mut pdus_by_id = HashMap::new();
     let mut batched_resolved_state = None;
     for pdus in &pdu_batches {
         batched_resolved_state = Some(
-            resolve_batch(&rules, pdus, &mut pdus_by_id, &mut batched_resolved_state)
-                .expect("batched state resolution step should succeed"),
+            resolve_batch(
+                &auth_rules,
+                &state_res_rules,
+                pdus,
+                &mut pdus_by_id,
+                &mut batched_resolved_state,
+            )
+            .expect("batched state resolution step should succeed"),
         );
     }
     let batched_resolved_state =
@@ -228,7 +241,8 @@ fn test_resolve(paths: &[&str]) -> Snapshots {
 
     // Resolve all PDUs in a single step
     let atomic_resolved_state = resolve_batch(
-        &rules,
+        &auth_rules,
+        &state_res_rules,
         pdu_batches.iter().flat_map(|x| x.iter()),
         &mut HashMap::new(),
         &mut None,
@@ -294,7 +308,8 @@ fn test_resolve(paths: &[&str]) -> Snapshots {
 ///
 /// # Arguments
 ///
-/// * `rules`: The rules of the room version.
+/// * `auth_rules`: The authorization rules of the room version.
+/// * `state_res_rules`: The state resolution rules of the room version.
 /// * `pdus`: An iterator of [`Pdu`]s to resolve, either alone or against the `prev_state`.
 /// * `pdus_by_id`: A map of [`OwnedEventId`]s to the [`Pdu`] with that ID.
 ///   * Should be empty for the first call.
@@ -303,7 +318,8 @@ fn test_resolve(paths: &[&str]) -> Snapshots {
 ///   * Should be `None` for the first call.
 ///   * Should not be mutated outside of this function.
 fn resolve_batch<'a, I, II>(
-    rules: &AuthorizationRules,
+    auth_rules: &AuthorizationRules,
+    state_res_rules: &StateResolutionV2Rules,
     pdus: II,
     pdus_by_id: &mut HashMap<OwnedEventId, Pdu>,
     prev_state: &mut Option<StateMap<OwnedEventId>>,
@@ -336,7 +352,10 @@ where
         auth_chain_sets.push(auth_events_dfs(&*pdus_by_id, pdu)?);
     }
 
-    resolve(rules, &state_sets, auth_chain_sets, |x| pdus_by_id.get(x).cloned()).map_err(Into::into)
+    resolve(auth_rules, state_res_rules, &state_sets, auth_chain_sets, |x| {
+        pdus_by_id.get(x).cloned()
+    })
+    .map_err(Into::into)
 }
 
 /// Perform state resolution on a batch of PDUs iteratively, one-by-one.
@@ -348,6 +367,7 @@ where
 /// # Arguments
 ///
 /// * `auth_rules`: The authorization rules of the room version.
+/// * `state_res_rules`: The state resolution rules of the room version.
 /// * `pdus`: An iterator of [`Pdu`]s to resolve, with the following assumptions:
 ///   * `prev_events` of each PDU points to another provided state event.
 ///
@@ -357,6 +377,7 @@ where
 /// to it via `prev_events`).
 fn resolve_iteratively<'a, I, II>(
     auth_rules: &AuthorizationRules,
+    state_res_rules: &StateResolutionV2Rules,
     pdus: II,
 ) -> Result<StateMap<OwnedEventId>, Box<dyn Error>>
 where
@@ -413,10 +434,13 @@ where
             auth_chains_before_event.push(auth_chain_at_event);
         }
 
-        let state_before_event =
-            resolve(auth_rules, &states_before_event, auth_chains_before_event.clone(), |x| {
-                pdus_by_id.get(x).cloned()
-            })?;
+        let state_before_event = resolve(
+            auth_rules,
+            state_res_rules,
+            &states_before_event,
+            auth_chains_before_event.clone(),
+            |x| pdus_by_id.get(x).cloned(),
+        )?;
 
         let auth_chain_before_event = auth_chain_from_state_map(&state_before_event)?;
 
@@ -434,6 +458,7 @@ where
 
         let state_at_event = resolve(
             auth_rules,
+            state_res_rules,
             &[state_before_event, proposed_state_at_event],
             vec![auth_chain_before_event, auth_chain_at_event],
             |x| pdus_by_id.get(x).cloned(),
@@ -468,8 +493,10 @@ where
         auth_chain_sets.push(auth_chain_at_event);
     }
 
-    resolve(auth_rules, &leaf_states, auth_chain_sets, |x| pdus_by_id.get(x).cloned())
-        .map_err(Into::into)
+    resolve(auth_rules, state_res_rules, &leaf_states, auth_chain_sets, |x| {
+        pdus_by_id.get(x).cloned()
+    })
+    .map_err(Into::into)
 }
 
 /// Depth-first search for the `auth_events` of the given PDU.